chore: improve TestConsensusEngineBuilder ergonomics (#3891)
